% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{aggregate_fragpipe_psm}
\alias{aggregate_fragpipe_psm}
\title{We groupby the fragpipe psm.tsv according to file and modified.peptide
We collapse by taking the row with maximum PeptideProphet.Probability or Probability.}
\usage{
aggregate_fragpipe_psm(df)
}
\arguments{
\item{df}{A FragPipe \code{psm.tsv} data frame to aggregate by raw file and modified peptide.}
}
\description{
We groupby the fragpipe psm.tsv according to file and modified.peptide
We collapse by taking the row with maximum PeptideProphet.Probability or Probability.
}
