% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{check_data_error}
\alias{check_data_error}
\title{na_policy determines how strict we allow NAs:
ignore: the columns can have any number of NA
allL: only if all values in column are NA, break operation
any: if even just one NA exist, break it.}
\usage{
check_data_error(
  data,
  required_cols = NULL,
  na_policy = c("any", "all", "ignore")
)
}
\description{
na_policy determines how strict we allow NAs:
ignore: the columns can have any number of NA
allL: only if all values in column are NA, break operation
any: if even just one NA exist, break it.
}
