% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{Colum_mapping_table}
\alias{Colum_mapping_table}
\title{This function is to generate the table for column mapping. i.e. which columns are used for what info}
\usage{
Colum_mapping_table(df, max_chars = 20)
}
\arguments{
\item{df}{data.frame to display}

\item{max_chars}{integer, maximum number of characters to show before truncating. Avoid wide columns}
}
\value{
data.frame with HTML \if{html}{\out{<span>}} for truncated display
}
\description{
This function is to generate the table for column mapping. i.e. which columns are used for what info
}
