% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpitoScope-package.R
\docType{package}
\name{EpitoScope-package}
\alias{EpitoScope}
\alias{EpitoScope-package}
\title{EpitoScope: Interactive QC, Visualization, and Immunogenicity Assessment of Immunopeptidomics Data}
\description{
An interactive R/Shiny application that unifies quality control, visualization, comparative statistics, and MHC class I/II binding prediction for mass-spectrometry-based immunopeptidomics data. It ingests search-engine outputs from PEAKS, FragPipe, DIA-NN, and Spectronaut through a configurable column-mapping layer, and provides publication-ready figure export and relational storage of processed results and binding predictions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yannic-chen/EpitoScope}
  \item Report bugs at \url{https://github.com/yannic-chen/EpitoScope/issues}
}

}
\author{
\strong{Maintainer}: Yannic Chen \email{yannicchn@gmail.com}

Authors:
\itemize{
  \item Yannic Chen \email{yannicchn@gmail.com}
}

}
\keyword{internal}
