% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{seed_db_once}
\alias{seed_db_once}
\title{Ensure the condition-terms table exists and is seeded (idempotent).}
\usage{
seed_db_once()
}
\description{
Ensure the condition-terms table exists and is seeded (idempotent).
}
\keyword{internal}
