% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{check_data_error}
\alias{check_data_error}
\title{na_policy determines how strict we allow NAs:
ignore: the columns can have any number of NA
allL: only if all values in column are NA, break operation
any: if even just one NA exist, break it.}
\usage{
check_data_error(
  data,
  required_cols = NULL,
  na_policy = c("any", "all", "ignore")
)
}
\arguments{
\item{data}{A data frame or a named list of data frames to validate.}

\item{required_cols}{A character vector of required column names.}

\item{na_policy}{How strictly to handle missing values: \code{"ignore"} allows any \code{NA},
\code{"all"} errors when all values in a required column are \code{NA}, and \code{"any"}
errors when any value in a required column is \code{NA}.}
}
\description{
na_policy determines how strict we allow NAs:
ignore: the columns can have any number of NA
allL: only if all values in column are NA, break operation
any: if even just one NA exist, break it.
}
